Reduce the multithreaded complexity of the code.
As ca8210_spi_transfer no longer happens in a context which cannot
sleep, it is simpler to use the synchonous sleepy versions of
functions rather than completions (which required further work to
be safe). This also simplifies a lot of the logic involved,
especially for retries, which were implemented in several different
locations.